pyproject.toml
[project]
name = "wpp-dental"
version = "0.1.0"
description = "Assistente IA de Agendamento via WhatsApp para Clínica Odontológica"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
